Enemy AI AttackRotation Fix + NewState (StopAttackIfPlayerDies)
In this update fix the character no longer cast a linetrace to its back, increased the range it can attack and added a Service in the blackboard logic to stop attacking the player if he is dead. Also Disabled Collisions on enemies so that when they die we no longer crash with their invisible mesh component.